Create Store Database (Games)
First, I created a table to introduce the games i have in the storefront. Then I introduced the games 1-15, making a table organized by the name, age of the customer, game rating, and price. Afterwards I selected the games where the rating was greater than 2.5 and added the cost of the filtered items together.